backend/pyproject.toml
[project]
name = "agentforge-backend"
version = "0.1.0"
description = "AgentForge API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "structlog>=24.4.0",
    "langgraph>=0.2.50",
    "langchain-core>=0.3.20",
    "redis[hiredis]>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "fakeredis[lua]>=2.23.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
